requirements.txt
# Core scientific computing
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0

# Stan interface
# Note: For conda users, prefer installing cmdstanpy via conda-forge:
#   conda install -c conda-forge cmdstanpy cmdstan
cmdstanpy>=1.2.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=0.19.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
tabulate>=0.8.9

# Testing
pytest>=7.0.0
